add commands to integration test with docker to makefile
pulls relevant environment variables to docker env from environment.sh
in the same process as other clients. Also removed a few unneccessary
verbose make commands and made the remaining commands more consistent
with usage in other apps. The ones you should now be interested in
(normally) are:

make test-with-docker
make integration-test-with-docker
make bash-with-docker