Require later version of JWT package
There have been two major releases since the version we were using. From
version `6.0.0` onwards the `JwtDecoder` takes a fourth argument, the
algorithm. There are no changes to how the code works, but we do need to
pass that argument in.

I've tested this change against the integration tests and it works with
versions 6.x and 7.x of JWT, so have let people use either instead of
pinning too precisely.