Upgrade dockerfile used for tests
We were using 2.1 of .NET core but this is not end of life.

You can see on https://hub.docker.com/_/microsoft-dotnet-sdk that
3.1 is long term support.

Also note, that the repo name has changed as per
https://github.com/dotnet/dotnet-docker/issues/2375.

Also note, I could have upgraded us to 5.1 but I went for 3.1 for two
reasons
- it's a smaller change
- good to know that we support the oldest .net stuff so we don't add
  anything in that is only supported by newer versions

However, I could be easily persuaded to choose 5.1 instead.