Downgrade docker image
In https://github.com/alphagov/notifications-net-client/pull/138 I
thought that 2.1 was no longer supported. This was incorrect. It is
supported until the end of August as per
https://dotnet.microsoft.com/platform/support/policy/dotnet-core

This, combined with the fact that actually if you run `make
test-with-docker` you will see that the change I added was broken and
the tests fail in docker.

Therefore by downgrading we support the lowest version possible and our
tests are fixed in docker. Win, Win.