migrate Notify to GovukNotify for v1.6.0
We're migrating this old package as it's still in use. This package is
pretty old, so tough to test these changes, but we've just done a
straight rename of the folders involved (and relevant lines in
makefiles, SLN, README.md, etc).

This is the last version of the .NET API Client to support .NET
Framework 4.5.2. This is used for building window apps, and is only
support up to windows 8.1/windows server 2012.

Windows 8.1 is on extended support until Q1 2023.
Windows Server 2012 is on extended suport until Q3 2023.