Stop rebuilding the image every time we run tests
Since we mount the local files into the image, we really don't need
to rebuild it every time we run the tests. On the other hand, we do
need to rebuild/recompile the code (otherwise why we are running the
tests?!), so this removes the image build in favour of recompiling.