Update description of Notify.Exceptions.NotifyClientException in docs
In most places we were saying that the client returns "a
`Notify.Exceptions.NotifyClientException` and an error code". It was
pointed out by a user that all the details returned when there is an
exception are actually in the message, so the docs were a little
unclear.

We already had one place where we the docs talked about the exception
"containing the relevant error code", so this uses that phrasing
everywhere.

It also changes the headings of the error table to be consistent, and
fixes a couple of places where we wrongly said the error that was
returned was an `HTTPError`.