Increase supported .NET Core version to 3.0
There are 3 incarnations of .NET:

- .NET Framework is the original Windows-only version of the .NET
runtime and class library.

- .NET Core was then created as a cross-platform version of .NET
and has differences to .NET Framework. It works in Docker.

- .NET Standard is a standard for class libraries that Framework
and Core both adhere to - to allow cross-implementation dev.

.NET Core v2.1 became EOL in August 2021 [^1] and there's no image
to run it on Mac M1 machines. We should bump in any case. Framework
and Standard are still comfortably in support* and changing the min
version of Core doesn't have any impact on those settings.

*.NET Standard is actually being deprecated with .NET Framework 5
[^2].

[^1]: https://dotnet.microsoft.com/en-us/platform/support/policy/dotnet-core
[^2]: https://devblogs.microsoft.com/dotnet/the-future-of-net-standard/