Bump target .NET (Core) version to 6
There are 3 incarnations of .NET:

- .NET Framework is the original Windows-only version of the .NET
runtime and class library.

- .NET Core was then created as a cross-platform version of .NET
and has differences to .NET Framework. It works in Docker.

- .NET Standard is a standard for class libraries that Framework
and Core both adhere to - to allow cross-implementation dev.

.NET 5 replaced .NET Core and .NET Standard with a single ".NET"
[^1] but is EOL in May [^2]. In addition, .NET Standard is still
supported so we should still target it.

What does the target version mean?
==================================

This represents the minimum required version for a particular .NET
incarnation. Example error for a .NET 6 package in .NET 5:

      Determining projects to restore...
      /foo/foo.csproj : error NU1202: Package Puya.Platform.Plugins.Auth.Shared 1.0.0.25 is not compatible with net5.0 (.NETCoreApp,Version=v5.0). Package Puya.Platform.Plugins.Auth.Shared 1.0.0.25 supports: net6.0 (.NETCoreApp,Version=v6.0)

Older versions of the package that target a lower version of .NET
can still be installed e.g. Serilog.Sinks.File 5.0.0 is a .NET 5
package that can still be installed in a .NET 6 environment.

[^1]: https://devblogs.microsoft.com/dotnet/the-future-of-net-standard/
[^2]: https://docs.microsoft.com/en-us/lifecycle/products/microsoft-net-framework