Extend maximum versions of JWT and NewtonSoft.JSON
The tests pass when manually pinned to the latest versions of both
of these libraries (8.9.0 and 13.0.1). Although we could just pin
the minimum version, the maximum is a useful indicator that we've
tested up to a certain point to avoid surprise incompatibilities.

The other packages are on their latest versions.