Fix release script not finding package to release
I'm not sure how this used to work: we were looking for the built
package in a place it could never be - unless the "dotnet pack -o"
behaviour has changed. I've tested this and the release now works.

Note to test you need to:

- Change the Appveyor project config to point at the appveyor.txt
for this branch [^1]. This is so Appveyor will pick up...

- Change "--branch=master" to "--branch=fix-publish-cmd-181753597"
in appveyor.txt. This is because the "release" stage of the pipeline
in Concourse doesn't tell Appveyor to build a specific "pullId" [^2].

[^1]: https://github.com/alphagov/notifications-net-client/pull/156
[^2]: https://github.com/alphagov/notifications-net-client/blob/258ea73a8ef96a90baaa7a4daa89dab226853936/appveyor.txt#L5