Remove unnecessary integration tests
An equivalent of
https://github.com/alphagov/notifications-php-client/pull/134/commits/576a35baf04895277de9a167a4bd90a9176ddabf
which means we don't need a new integration tests for every
parameter our methods have and we instead test all of them together
(we already have this test) and treat that as sufficient.
There are probably some more integration tests that could go
based on this but I've started small to be less controversial.