Format the repo with dotnet format
Also adds helpful commands for others to run this in the future
if they choose. I didn't build this into the requirements for the
code to pass its tests as I'm not too familiar with .net and this
is still a starting improvement.

dotnet format is built into .net 6 SDK by default so no new
packages or controversial choices here - just using the default
thing dotnet give us.

https://github.com/dotnet/format