Update integration tests to wait before checking SMS cost data
The tests which check if the `costDetails` properties for an SMS
notification are not null have been failing at times because those
fields are null. Those properties aren't ready immediately, but we can
check if they are by checking whether `isCostDataReady` is true and
adding in a wait if the cost data is not ready yet.